+++ Studio/src/Application/Analysis/AnalysisTool.cpp
@@ -8,8 +8,8 @@
 
 // shapeworks
 #include <Analysis/AnalysisTool.h>
-#include <Visualization/ShapeWorksStudioApp.h>
-#include <Visualization/ShapeWorksWorker.h>
+#include <Interface/ShapeWorksStudioApp.h>
+#include <Data/ShapeWorksWorker.h>
 #include <Data/Session.h>
 #include <Data/StudioMesh.h>
 #include <Data/Shape.h>
