+++ Studio/src/Application/Interface/ShapeWorksStudioApp.h
