+++ Studio/src/Application/main.cpp
@@ -3,7 +3,7 @@
 #include <QMessageBox>
 #include <QDir>
 
-#include <Visualization/ShapeWorksStudioApp.h>
+#include <Interface/ShapeWorksStudioApp.h>
 #include <iostream>
 
 #include <vtkObject.h>
